requirements.txt
alembic==1.14.0
asttokens==3.0.0
# autofeat==2.1.3
Boruta==0.4.3
build==1.2.2.post1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
docutils==0.21.2
executing==2.1.0
FeatureFlex==0.1.12
filelock==3.16.1
flexcache==0.3
flexparser==0.4
fonttools==4.55.3
fsspec==2024.12.0
greenlet==3.1.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.31.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kagglehub==0.3.6
keyring==25.6.0
kiwisolver==1.4.8
lightgbm==4.5.0
llvmlite==0.43.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.20
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
optuna==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfkit==1.0.0
pexpect==4.9.0
pillow==11.0.0
Pint==0.24.4
pkginfo==1.12.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
scikit-learn==1.5.0
scipy==1.14.1
SecretStorage==3.3.3
setuptools==75.6.0
shap==0.46.0
six==1.17.0
sklearn-compat==0.1.3
skrebate==0.62
slicer==0.0.8
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
twine==6.0.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
xgboost==2.1.3
